fix OD and OA symbols code in first line of MB registry file

+++ eirc/src/org/flexpay/eirc/sp/impl/MbParsingConstants.java
@@ -9,5 +9,5 @@ public abstract class MbParsingConstants {
 			"                                                                                                    "
 			+ "                                                                                                    "
 			+ "                                                                                                    ";
-	public static final long FIRST_FILE_STRING_SIZE = 300;
+	public static final int FIRST_FILE_STRING_SIZE = 300;
 }
